Refactor PDF export with utility functions and chart data processing
- Extract chart-related and PDF export logic into separate utility functions
- Move chart generation, data processing, and PDF export to dedicated utility modules
- Simplify PDF export method by leveraging new utility functions
- Improve code organization and readability
